creating histograms in a background task. running image edits in separate thread in the background. adding (atomic) progress reporting to long image editing operations. adding (atomic) fixed point class. used for progress reporting instead of a float, since there is not std::atomic<float> using ImageButton as progress bar during long operations. changed image modification API to return an HDRImage instead of modifying the current one (needed to avoid race conditions with background modify tasks). using shared_ptrs instead of raw pointers to store images and async tasks. adding separate callbacks to HDRImageManager when image modification tasks start and finish. adding some commonly typedefs throughout to clean up the code a bit. updating ImageButton info (caption, isModified) without repopulating entire image list.

+++ src/hdrimageio.cpp
@@ -275,7 +275,7 @@ bool HDRImage::load(const string & filename)
 		size_t imageIndex = size_t(-1);
 		{
 			size_t largest = 0;
-			size_t largestWidth = images[0].width;
+			int largestWidth = images[0].width;
 			for (size_t i = 0; i < images.size(); i++)
 			{
 				if (largestWidth < images[i].width)
